Implement Listener::FTP, fully functional without notification.
The FTP listener listens to one or more directories in an FTP account,
looking for new files. It stores filenames in Redis to keep track of
what's been seen.

Checksums are constructed and stored based on the mtime and size of the
file. These are available over FTP without magic. The purpose of this
tool is to watch for changes (new or altered files), for which the
checksum will likely be sufficient. If proper hashes are needed, this
can probably be left to worker-level (i.e. outside of this tool).