Introduce config --redis-ns-notifiers, for namespacing Redis notifiers.
The default, which was previously hard-coded, of `resque:` is retained.
This is normally used for Resque, and works with Sidekiq. Sidekiq,
however, allows more flexibility in namespacing.

Note that it is not currently possible to use different notifier
namespaces for `Notifier::Sidekiq`, despite it being possible to use
multiple `--notifier-sidekiq` parameters pointing to different Redis
URLs.