Fix Listener::FTP missing port throughout.
Previously, it was not possible to configure `Listener::FTP` port
(besides using `FTP_PORT` environment variable to pass through to
`Net::FTP`), and the port was not part of `#redis_key`. There's no
reason why multiple FTP servers can't be run on different ports, so
this change is needed for completeness. Add `--ftp-port` option to
`bin/ryespy`.

Note that this change is not backwards-compatible! Sorry. :( This is
intended for the same release as 22f4ce7, however, and we haven't yet
reached the first major version, so hopefully this won't cause too much
disappointment.