Rename -v => -d , fixing App#refresh_loop to reraise if DEBUG.
Verbose mode `-v` is actually better described as debug mode `-d`, as
it also explodes with stack traces if enabled. It is intended that the
option is not used generally, so transient errors will be ignored and
the operations retried when running with `--eternal`. Also fix
`App#refresh_loop` to reraise stack traces when running in debug mode.

When using the `bin/ryespy` executable in development, you'll probably
want to set this `--debug`,`-d` mode.