Implement Listener::GoogCS , for Google Cloud Storage.
The `GoogCS` listener listens to one or more file key prefix filters in
a Google Cloud Storage bucket, looking for new or changed files.

The implementation is almost identical to `Listener::AmznS3`. At least
for now, the listeners are handled fairly separately, however, to allow
for any service-specific tweaks. c.f. `Listener::RaxCF`, which needs
a specific treatment for authentication endpoints and regions.

`README` and `ryespy.gemspec` are updated to document usage and the
extended utility of Ryespy.

This listener is fully functional.