Extend Listener::GoogDrv#get_unseen_files to detect changes.
Previously, `Listener::GoogDrv#get_unseen_files` detected new files.
Extend to read `updated`, with `etag` if present, and use these as a
checksum. This is similar to `Listener::FTP#get_unseen_files`.

Note that this might cause notifications to be made even if the content
has not changed, but if that is important, it is expected that it will
be handled and filtered worker-side.